merge

+++ src/Controller/EventController.php
@@ -45,6 +45,7 @@ class EventController extends AbstractController
     )]
     public function index(Request $request): Response
     {
+        $owner = $this->getUser();
         $page = $request->query->getInt('page', 1);
 
         $categoryId = $request->query->get('categoryId');
@@ -56,6 +57,7 @@ class EventController extends AbstractController
         $title = $request->query->get('title');
 
         $pagination = $this->eventService->getPaginatedList(
+            $owner,
             $page,
             $categoryId,
             $title,
@@ -86,7 +88,9 @@ class EventController extends AbstractController
     )]
     public function new(Request $request): Response
     {
+        $user = $this->getUser();
         $event = new Event();
+        $event->setOwner($user);
 
         $form = $this->createForm(EventType::class, $event);
         $form->handleRequest($request);
@@ -129,31 +133,6 @@ class EventController extends AbstractController
     }
 
     /**
-     * Show action.
-     *
-     * @param Event $event event
-     *
-     * @return Response HTTP response
-     */
-    #[Route(
-        '/{id}',
-        name: 'app_event_show',
-        requirements: ['id' => '[1-9]\d*'],
-        methods: ['GET']
-    )]
-    public function show(Event $event): Response
-    {
-        $this->denyAccessUnlessGranted(
-            EventVoter::VIEW,
-            $event
-        );
-
-        return $this->render('event/show.html.twig', [
-            'event' => $event,
-        ]);
-    }
-
-    /**
      * Edit action.
      *
      * @param Request $request request
@@ -167,8 +146,17 @@ class EventController extends AbstractController
         requirements: ['id' => '[1-9]\d*'],
         methods: ['GET', 'PUT'],
     )]
+    #[IsGranted(EventVoter::EDIT, 'event')]
     public function edit(Request $request, Event $event): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') && $event->getOwner() !== $this->getUser()) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.record_not_found')
+            );
+
+            return $this->redirectToRoute('app_event_index');
+        }
         $form = $this->createForm(
             EventType::class,
             $event,
@@ -239,8 +227,17 @@ class EventController extends AbstractController
         requirements: ['id' => '[1-9]\d*'],
         methods: ['GET', 'DELETE'],
     )]
+    #[IsGranted(EventVoter::DELETE, 'event')]
     public function delete(Request $request, Event $event): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') && $event->getOwner() !== $this->getUser()) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.record_not_found')
+            );
+
+            return $this->redirectToRoute('app_event_index');
+        }
         $form = $this->createForm(EventType::class, $event, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('app_event_delete', ['id' => $event->getId()]),
